+++ spec/integration/membership_spec.rb
@@ -21,12 +21,78 @@ describe 'request to a protected resource' do
 
         context 'and not team member' do
           before { github_login }
-          it { is_expected.to be_not_found}
+          it { is_expected.to be_not_found }
+        end
+      end
+    end
+
+    context 'which is specified by multiple numeric team ids' do
+      subject { get '/multi_team/protected' }
+
+      context 'when not logged in' do
+        it { is_expected.to be_github_oauth_redirect }
+      end
+
+      context 'when logged in' do
+        context 'and a first team member' do
+          before do
+            user = github_login
+            user.stub_membership(team: 123)
+          end
+
+          it { is_expected.to be_ok }
+        end
+
+        context 'and another team member' do
+          before do
+            user = github_login
+            user.stub_membership(team: 345)
+          end
+
+          it { is_expected.to be_ok }
+        end
+
+        context 'and not team member' do
+          before { github_login }
+          it { is_expected.to be_not_found }
         end
       end
     end
 
     context 'which is specified by a team alias' do
+      subject { get '/multi_team_alias/protected' }
+
+      context 'when not logged in' do
+        it { is_expected.to be_github_oauth_redirect }
+      end
+
+      context 'when logged in' do
+        context 'and a first team member' do
+          before do
+            user = github_login
+            user.stub_membership(team: 456)
+          end
+
+          it { is_expected.to be_ok }
+        end
+
+        context 'and another team member' do
+          before do
+            user = github_login
+            user.stub_membership(team: 789)
+          end
+
+          it { is_expected.to be_ok }
+        end
+
+        context 'and not team member' do
+          before { github_login }
+          it { is_expected.to be_not_found }
+        end
+      end
+    end
+
+    context 'which is specified by multiple team aliases' do
       subject { get '/team_alias/protected' }
 
       context 'when not logged in' do
@@ -45,7 +111,7 @@ describe 'request to a protected resource' do
 
         context 'and not team member' do
           before { github_login }
-          it { is_expected.to be_not_found}
+          it { is_expected.to be_not_found }
         end
       end
     end
@@ -65,7 +131,7 @@ describe 'request to a protected resource' do
 
         context 'and not team member' do
           before { github_login }
-          it { is_expected.to be_not_found}
+          it { is_expected.to be_not_found }
         end
       end
     end
@@ -113,7 +179,7 @@ describe 'request to a protected resource' do
 
         context 'and not organization member' do
           before { github_login }
-          it { is_expected.to be_not_found}
+          it { is_expected.to be_not_found }
         end
       end
     end
@@ -136,7 +202,34 @@ describe 'request to a resource that only exists when logged in' do
 
       context 'when not team member' do
         before { github_login }
-        it { is_expected.to be_not_found}
+        it { is_expected.to be_not_found }
+      end
+    end
+
+    context 'which is specified by multiple numeric team ids' do
+      subject { get '/multi_team/conditional' }
+
+      context 'when a first team member' do
+        before do
+          user = github_login
+          user.stub_membership(team: 123)
+        end
+
+        it { is_expected.to be_ok }
+      end
+
+      context 'when another team member' do
+        before do
+          user = github_login
+          user.stub_membership(team: 345)
+        end
+
+        it { is_expected.to be_ok }
+      end
+
+      context 'when not team member' do
+        before { github_login }
+        it { is_expected.to be_not_found }
       end
     end
 
@@ -154,7 +247,34 @@ describe 'request to a resource that only exists when logged in' do
 
       context 'when not team member' do
         before { github_login }
-        it { is_expected.to be_not_found}
+        it { is_expected.to be_not_found }
+      end
+    end
+
+    context 'which is specified by multiple team aliases' do
+      subject { get '/multi_team_alias/conditional' }
+
+      context 'when a member of the first team' do
+        before do
+          user = github_login
+          user.stub_membership(team: 456)
+        end
+
+        it { is_expected.to be_ok }
+      end
+
+      context 'when a member of another team' do
+        before do
+          user = github_login
+          user.stub_membership(team: 789)
+        end
+
+        it { is_expected.to be_ok }
+      end
+
+      context 'when not a team member' do
+        before { github_login }
+        it { is_expected.to be_not_found }
       end
     end
   end