+++ spec/rails_app/config/routes.rb
@@ -17,9 +17,13 @@ RailsApp::Application.routes.draw do
 
   github_authenticate(team: 123)  { get '/team/protected'   => responses[200] }
   github_authenticated(team: 123) { get '/team/conditional' => responses[200] }
+  github_authenticate(team: [123, 345])  { get '/multi_team/protected'   => responses[200] }
+  github_authenticated(team: [123, 345]) { get '/multi_team/conditional' => responses[200] }
 
   github_authenticate(team: :marketing)  { get '/team_alias/protected'   => responses[200] }
   github_authenticated(team: :marketing) { get '/team_alias/conditional' => responses[200] }
+  github_authenticate(team: [:marketing, :interns])  { get '/multi_team_alias/protected'   => responses[200] }
+  github_authenticated(team: [:marketing, :interns]) { get '/multi_team_alias/conditional' => responses[200] }
 
   github_authenticate(org: :foobar_inc)  { get '/org/protected'   => responses[200] }
   github_authenticated(org: :foobar_inc) { get '/org/conditional' => responses[200] }